+++ src/user/use-cases/list-user-by-id/list-user-by-id.spec.ts
@@ -1,6 +1,6 @@
 import { NotFoundException } from '@nestjs/common';
 import { User } from '@/user/infra/entities';
-import { ListUserUseCase } from './list-user';
+import { ListUserUseCase } from './list-user-by-id';
 import { IUserRepository } from '@/user/infra/interfaces';
 
 describe('Get User', () => {
